Refactor work ongoing, add sample data
This commit continues ongoing refactor work for the terribly-named pubsub branch. JSON is outputting now. Next step is to pass that JSON into a database -- likely RethinkDB -- to allow workers to grab it and make sense of it for the public.

I've also started a library of sample data -- first a Dense Fog Advisory.